+++ app/__init__.py
@@ -1,3 +1,4 @@
-from .routes import router as file_router
+from .file_routes import file_router
+from .folder_routes import folder_router
 
-__all__ = ['file_router']
+__all__ = ['file_router', 'folder_router']
